// Frontend/src/app/auth/models/RegistrationCredentials.ts
export class RegistrationCredentials {
  firstName: string | undefined;
  secondName: string | undefined;
  thirdName: string | undefined;
  birthDate: Date | undefined;
  phone: string | undefined;
  email: string | undefined;
  userName: string | undefined;
  password: string | undefined;

constructor()  {
}

}